Let slides scale height based on content
By setting min-height instead of height to the slide, and using relative
positioning for the current one, let the slide scale its height based on
content.

To get around a graphical glitch when navigating vertically to the
previous slide when using transitions and relative positioning, force
all absolutely positioned slides to align to the top.